Switch iOS IPA target from TF2 to offline CS:GO without Steam or Scaleform.

+++ game/client/imessagechars.h
@@ -1,4 +1,4 @@
-//========= Copyright Valve Corporation, All rights reserved. ============//
+//========= Copyright © 1996-2005, Valve Corporation, All rights reserved. ============//
 //
 // Purpose: 
 //
@@ -16,12 +16,12 @@
 #pragma once
 #endif
 
-#include <vgui/VGUI.h>
+#include <vgui/vgui.h>
 
 namespace vgui
 {
 class Panel;
-typedef uint32 HFont;
+typedef unsigned long HFont;
 }
 
 abstract_class IMessageChars
